set selected file to undefined on conflict & move conflict service to its own file
